src/sensors/ultrasonic/requirements.txt
paho-mqtt
RPI.GPIO